Split typechecker file-size guards
