+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -1,3 +1,4 @@
 mod focused_tests;
 mod lexer_and_monomorphize;
 mod thresholds;
+mod typechecker_test_splits;
